requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3

reportlab~=4.4.1